xenstore index

Library xenstore

This library exposes the following toplevel modules:

Xs_handle

Xs_protocol

Library xenstore.client

The entry point of this library is the module: Xs_client_lwt.

Library xenstore.server

The entry point of this library is the module: Xenstore_server.

Library xenstore.unix

The entry point of this library is the module: Xs_client_unix.